requirements.txt
pytelegrambotapi==4.16.0
gspread==5.12.0
oauth2client==4.1.3
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0